+++ demo/public/src/app/models/i-task.d.ts
@@ -1,3 +1,0 @@
-export interface ISay {
-    text: string;
-}