+++ demo/public/src/app/models/i-task.js
@@ -1,3 +1,0 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-//# sourceMappingURL=data:application/json;base64,eyJ2ZXJzaW9uIjozLCJmaWxlIjoiaXNheS5qcyIsInNvdXJjZVJvb3QiOiJDOi9Db2RlL0BrYWxlYWJtZWxraWUvb3VyLXNheS8iLCJzb3VyY2VzIjpbInB1YmxpYy9zcmMvYXBwL21vZGVscy9pc2F5LnRzIl0sIm5hbWVzIjpbXSwibWFwcGluZ3MiOiIifQ==