+++ demo/public/src/app/models/i-task.ts
@@ -1,7 +1,0 @@
-export interface ITask {
-  _id?: string
-  done: boolean
-  created: number
-  title: string
-  desc?: string
-}