+++ demo/src/routers/api/tasks.router.d.ts
@@ -1,4 +1,0 @@
-/// <reference types="koa-router" />
-import * as Router from 'koa-router';
-declare const TasksRouter: Router;
-export { TasksRouter };