+++ demo/src/sockets/api/tasks.socket.d.ts
@@ -1,3 +1,0 @@
-/// <reference types="socket.io" />
-declare const TasksSocket: SocketIO.Server;
-export { TasksSocket };